schema_mapper/requirements.txt
google-adk
google-cloud-bigquery
google-cloud-storage
functions-framework
